Moved "tests for tests" to a separate directory.
